refactor: extract shared OKLink API key helper for xlayer adapters
