+++ active-users/xlayer.ts
@@ -1,21 +1,11 @@
 import { FetchOptions, ProtocolType, SimpleAdapter } from "../adapters/types";
 import { CHAIN } from "../helpers/chains";
-import { getEnv } from "../helpers/env";
+import { getOklinkApiKey } from "../helpers/oklink";
 import { httpGet } from "../utils/fetchURL";
-
-async function getApiKey(): Promise<string> {
-  const API_KEY = getEnv('OKLINK_API_KEY');
-  if (!API_KEY) throw Error('Missing env OKLINK_API_KEY');
-  const s = 1111111111111;
-  const rotated = `${API_KEY.slice(8)}${API_KEY.slice(0, 8)}`;
-  const now = Date.now();
-  const time = `${(now + s).toString()}${Math.floor(Math.random() * 1000).toString().padStart(3, '0')}`;
-  return Buffer.from(`${rotated}|${time}`).toString('base64');
-}
 
 const fetch = async (options: FetchOptions) => {
   const timestamp = options.startOfDay * 1e3;
-  const apiKey = await getApiKey();
+  const apiKey = await getOklinkApiKey();
   const headers = { 'x-apikey': apiKey };
   const [activeData, txData, gasData] = await Promise.all([
     httpGet(`https://www.oklink.com/api/explorer/v2/common/charts/activeAddressCount?chain=X1&t=${timestamp}`, { headers }),
